+++ bzt/cli.py
@@ -30,7 +30,7 @@ from colorlog import ColoredFormatter
 import bzt
 from bzt import ManualShutdown, NormalShutdown, RCProvider, AutomatedShutdown
 from bzt.engine import Engine, Configuration, ScenarioExecutor
-from bzt.six import HTTPError, string_types, b
+from bzt.six import HTTPError, string_types, b, get_stacktrace
 from bzt.utils import run_once, is_int, BetterDict, is_windows, is_piped
 
 
@@ -51,6 +51,7 @@ class CLI(object):
         self.log.debug("Python: %s %s", platform.python_implementation(), platform.python_version())
         self.log.debug("OS: %s", platform.uname())
         self.engine = Engine(self.log)
+        self.exit_code = 0
 
     @staticmethod
     @run_once
@@ -148,7 +149,6 @@ class CLI(object):
         :type configs: list
         :return: integer exit code
         """
-        overrides = []
         jmx_shorthands = []
         try:
             jmx_shorthands = self.__get_jmx_shorthands(configs)
@@ -157,47 +157,49 @@ class CLI(object):
             self.__configure(configs)
             self.engine.prepare()
             self.engine.run()
-            exit_code = 0
         except BaseException as exc:
-            self.log.debug("Caught exception in try: %s", traceback.format_exc())
-            if isinstance(exc, ManualShutdown):
-                self.log.info("Interrupted by user: %s", exc)
-            elif isinstance(exc, NormalShutdown):
-                self.log.info("Normal shutdown")
-            elif isinstance(exc, AutomatedShutdown):
-                self.log.info("Automated shutdown")
-            else:
-                if isinstance(exc, HTTPError):
-                    self.log.warning("Response from %s: %s", exc.geturl(), exc.read())
-                self.log.error("%s: %s", type(exc).__name__, exc)
-            exit_code = 1
+            self.handle_exception(exc)
         finally:
             try:
-                for fname in overrides + jmx_shorthands:
+                for fname in jmx_shorthands:
                     os.remove(fname)
                 self.engine.post_process()
-            except KeyboardInterrupt as exc:
-                self.log.debug("Exception: %s", traceback.format_exc())
-                exit_code = 1
-                if isinstance(exc, RCProvider):
-                    exit_code = exc.get_rc()
             except BaseException as exc:
-                self.log.debug("Caught exception in finally: %s", traceback.format_exc())
-                self.log.error("%s: %s", type(exc).__name__, exc)
-                exit_code = 1
-
-        if isinstance(self.engine.stopping_reason, RCProvider):
-            exit_code = self.engine.stopping_reason.get_rc()
+                self.handle_exception(exc)
 
         self.log.info("Artifacts dir: %s", self.engine.artifacts_dir)
-        if exit_code:
-            self.log.warning("Done performing with code: %s", exit_code)
+
+        if self.exit_code:
+            self.log.warning("Done performing with code: %s", self.exit_code)
         else:
-            self.log.info("Done performing with code: %s", exit_code)
+            self.log.info("Done performing with code: %s", self.exit_code)
 
         self.__close_log()
 
-        return exit_code
+        return self.exit_code
+
+    def handle_exception(self, exc):
+        info_level = http_level = default_level = logging.DEBUG
+        if not self.exit_code:  # only fist exception goes to the screen
+            info_level = logging.WARNING
+            http_level = logging.WARNING
+            default_level = logging.ERROR
+            if isinstance(exc, RCProvider):
+                self.exit_code = exc.get_rc()
+            else:
+                self.exit_code = 1
+
+        if isinstance(exc, ManualShutdown):
+            self.log.log(info_level, "Interrupted by user")
+        elif isinstance(exc, AutomatedShutdown):
+            self.log.log(info_level, "Automated shutdown")
+        elif isinstance(exc, NormalShutdown):
+            self.log.log(info_level, "Normal shutdown")
+        elif isinstance(exc, HTTPError):
+            self.log.log(http_level, "Response from %s: %s", exc.geturl(), exc.read())
+        else:
+            self.log.log(default_level, "%s: %s", type(exc).__name__, exc)
+            self.log.log(default_level, get_stacktrace(exc))
 
     def __get_jmx_shorthands(self, configs):
         """
@@ -205,7 +207,6 @@ class CLI(object):
         :type configs: list
         :return: list
         """
-
         jmxes = []
         for filename in configs[:]:
             if filename.lower().endswith(".jmx"):